requirements.txt
bcrypt==3.2.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
Flask==1.0.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
pycparser==2.20
python-decouple==3.3
python-dotenv==0.13.0
requests==2.19.1
six==1.15.0
urllib3==1.23
Werkzeug==1.0.1
